Правки к 3 лабе
